server/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlalchemy[asyncio]>=2.0.36
aiosqlite==0.20.0
aiofiles>=23.2.1
asyncpg==0.29.0  # PostgreSQL driver for Cloud SQL
python-jose[cryptography]==3.3.0
bcrypt==4.1.2
alembic==1.13.1
python-multipart==0.0.9
python-dotenv==1.0.1
pydantic-settings==2.2.1
